requirements.txt
flask
requests
img2pdf
pyyaml
mwoauth
mwclient
